[BUG]: home,search controller 로그인, 비로그인 로직 분리
